% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentrate.R
\name{concentrate}
\alias{concentrate}
\title{Concentrate observations at geographic coordinates.}
\usage{
concentrate(.data, lambda = 0.05, lat = lat, long = long,
  lat_new = NULL, long_new = NULL, group_var = NULL, ...)
}
\arguments{
\item{.data}{A tbl with latitude and longitude}

\item{lambda}{A clustering coefficient. Used in combination with the total
range spanned by the data to estimate a cluster cutoff.}

\item{lat, long}{Column names (unquoted) or positions of latitude and
longitude in `.data`}

\item{lat_new, long_new}{Column names (quoted) for storing the projection
results. If NULL, the original latitude/longitude columns are used.}

\item{group_var}{if not-`NULL` keep a grouping variable for the concentrated
locations with this column name}

\item{...}{Currently not used.}
}
\description{
Concentrate observations that are close to each other at a few representative
locations. This is useful primarily for displaying purposes. It allows to
declutter observations and reduce locally scattered groups to a few locations
of interest.
}
\examples{
library(tidyverse)

df <- tibble(
  foo = rep(LETTERS[1:3], each = 3),
  lat = rnorm(9, rep(c(-30, -20, 50), each = 3)),
  long = rnorm(9, rep(c(-90, -100, 85), each = 3)))

df \%>\% concentrate()

df \%>\% concentrate(.2, long_new = 'x', lat_new = 'y', group_var = "i")
}
