% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_graticule.R
\name{geom_graticule}
\alias{geom_graticule}
\title{Graticules}
\usage{
geom_graticule(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", proj = "longlat", long_0 = 0, lat_min = -90,
  lat_max = 90, lat_by = 20, lat_n = NULL, long_min = -180 + long_0,
  long_max = 180 + long_0, long_by = 30, long_n = NULL,
  proj_extra = NULL, ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = FALSE)
}
\description{
grid lines for maps
}
\details{
inspired by https://gist.github.com/valentinitnelav/c7598fcfc8e53658f66feea9d3bafb40
}
