% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{Project geographic coordinates}
\usage{
project(.data, proj = NULL, long_0 = NULL, scale_factor = 1e-05,
  lat = lat, long = long, lat_new = NULL, long_new = NULL, ...)
}
\arguments{
\item{.data}{A tbl with latitude and longitude}

\item{proj}{A name of a cartographic projection or a complete proj4-string
specifying a transformation supported by the \href{http://proj4.org}{PROJ.4}
project.}

\item{long_0}{Latitude of the central meridian (added to the proj4-string as
`+lon_0=<long_0>`}

\item{scale_factor}{to apply to results of projections. PROJ4 uses meters as
resolution for most transformation.}

\item{lat, long}{Column names (unquoted) or positions of latitude and
longitude in `.data`}

\item{lat_new, long_new}{Column names (quoted) for storing the projection
results. If NULL, the original latitude/longitude columns are used.}

\item{...}{other arguments passed to [rgdal::project]}
}
\description{
`project()` is a wrapper around [rgdal::project()] that plays nice with
magrittr/tidyverse-style pipes and variable selection syntax.
}
\examples{
library(tidyverse)

tibble(lat = c(-45,0,45), long = 60, stuff = LETTERS[1:3]) \%>\%
  project("robin", 90)

tibble(stuff = LETTERS[1:3], LAT = c(-45,0,45), LON = 60) \%>\%
  project("robin", 90, 1, LAT, LON, "lat robin [m]", "long robin [m]")
}
