% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_worldmap}
\alias{theme_worldmap}
\alias{theme_worldmap_light}
\title{Worldmap themes}
\usage{
theme_worldmap_light(...)
}
\description{
Themes based on ggplot2 themes that play well with ggworldmaps.
}
\details{
\describe{

\item{`theme_gray`}{A theme based on ‘ggplot2::theme_light’, with light grey
lines and axes, to direct more attention towards the data.}

}
}
