% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_worldmap.R
\name{worldmap}
\alias{worldmap}
\title{Worldmap Polygons}
\usage{
worldmap(map = "world", proj = NULL, long_0 = 0, long_min = -180 +
  long_0, long_max = 180 + long_0, lat_min = -80, lat_max = 80,
  fill = TRUE, plot = FALSE, ...)
}
\arguments{
\item{proj}{A name of a cartographic projection or a complete proj4-string
specifying a transformation supported by the \href{http://proj4.org}{PROJ.4}
project.}

\item{long_0}{Latitude of the central meridian (added to the proj4-string as
`+lon_0=<long_0>`}

\item{long_min, long_max}{Longitude limits, defaults to ±180° from long_0.}

\item{lat_min, lat_max}{Latitude limits, defaults to ±80°.}

\item{...}{other arguments passed to [rgdal::project]}
}
\description{
Worldmap Polygons
}
