% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_degrees.R, R/geom_graticule.R
\name{geom_degree}
\alias{geom_degree}
\alias{degrees}
\alias{geom_graticule}
\alias{geom_gratframe}
\alias{graticules}
\title{Graticules and Degree Labels}
\usage{
geom_degree(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", proj = NULL, long_0 = 0, lat_min = -80,
  lat_max = 80, lat_by = 20, lat_n = NULL, lat_at = c(long_min,
  long_max), lat_nudge_x = 5, lat_nudge_y = 0, lat_vjust = "center",
  lat_hjust = "outward", long_min = -180 + long_0, long_max = 180 +
  long_0, long_by = 40, long_n = NULL, long_at = c(lat_min, lat_max),
  long_nudge_x = 0, long_nudge_y = 5, long_vjust = "outward",
  long_hjust = "center", ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = FALSE)

degrees(proj = NULL, long_0 = 0, lat_min = -80, lat_max = 80,
  lat_by = 20, lat_n = NULL, lat_at = c(long_min, long_max),
  lat_nudge_x = 5, lat_nudge_y = 0, long_min = -180 + long_0,
  long_max = 180 + long_0, long_by = 40, long_n = NULL,
  long_at = c(lat_min, lat_max), long_nudge_x = 0, long_nudge_y = 10)

geom_graticule(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", proj = NULL, long_0 = 0, lat_min = -80,
  lat_max = 80, lat_by = 20, lat_n = NULL, long_min = -180 + long_0,
  long_max = 180 + long_0, long_by = 40, long_n = NULL, ...,
  na.rm = TRUE, show.legend = NA, inherit.aes = FALSE)

geom_gratframe(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", proj = NULL, long_0 = 0, lat_min = -80,
  lat_max = 80, long_min = -180 + long_0, long_max = 180 + long_0, ...,
  na.rm = TRUE, show.legend = NA, inherit.aes = FALSE)

graticules(proj = NULL, long_0 = 0, lat_min = -80, lat_max = 80,
  lat_by = 20, lat_n = NULL, long_min = -180 + long_0, long_max = 180 +
  long_0, long_by = 40, long_n = NULL)
}
\arguments{
\item{data}{If `NULL` uses [graticules()] and [degrees()] to compute the
required data.}
}
\description{
Add graticules, degree labels and a frame to a map plot.
}
\details{
\describe{
  \item{Graticules}{inspired by \url{https://gist.github.com/valentinitnelav/c7598fcfc8e53658f66feea9d3bafb40}}
}
}
