% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggworldmap.R
\name{ggworldmap}
\alias{ggworldmap}
\title{Create a ggworldmap with graticules and degrees}
\usage{
ggworldmap(data = NULL, mapping = aes(x = long, y = lat), map = "world",
  proj = NULL, long_0 = 0, long_min = -180 + long_0, long_max = 180 +
  long_0, lat_min = -84, lat_max = 84, proj_extra = NULL,
  style = c("light"), ..., graticule = style, gratframe = style,
  degree = style, theme = style, environment = parent.frame())
}
\arguments{
\item{data}{Default dataset to use for plot. If not already a data.frame,
will be converted to one by \code{\link[=fortify]{fortify()}}. If not specified,
must be supplied in each layer added to the plot.}

\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{style}{Default style for the map, graticules, degrees and the default
theme. Choices: "light".}

\item{...}{Other arguments passed on to [geom_worldmap]}

\item{graticule}{Style and arguments passed on to [geom_graticule]. Either a
single style string, a list of args with the style as first argument, or NULL
to disable graticules.}

\item{gratframe}{Style and arguments passed on to [geom_gratframe]. Either a
single style string, a list of args with the style as first argument, or NULL
to disable a gratiframe.}

\item{degree}{Style and arguments passed on to [geom_degree]. Either a single
style string, a list of args with the style as first argument, or NULL to
disable degrees.}

\item{theme}{Style and arguments passed on to `theme_worldmap_<style>`
themes. Either a single style string, a list of args with the style as first
argument, or NULL to disable a default theme.}

\item{environment}{If a variable defined in the aesthetic mapping is not
found in the data, ggplot will look for it in this environment. It defaults
to using the environment in which \code{ggplot()} is called.}
}
\value{
ggplot object
}
\description{
`ggworldmap()` is a short cut to create a [ggplot2::ggplot] with polygons of
a world map polygons, graticules and degree labels. It also adds a customized
theme and a fixed aspect ratio coordinate system ([ggplot2::coord_equal]).
}
\details{
`lat_min/lat_max` default to 84°S/W, respectively. This is mainly because the
polygon map that comes with ggplot2 only covers that range. Also, in most
projections, those extreme latitudes are heavily distorted and plotting them
doesn't make sense anyway.
}
\examples{
ggworldmap()

ggworldmap(proj = "robin", degree = NULL)

ggworldmap(proj = "merc", long_0 = -90, lat_min = -75, lat_max = 75,
  mapping = aes(fill=region), color = "black",
  graticule = list("light", color = "grey50"))
}
